requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
python-dotenv==0.19.0